requirements.txt
streamlit
tensorflow
numpy
Pillow
lime
shap
matplotlib 
scikit-image 
seaborn
